It really is alive!
It still differs from the official specification (arguments are passed
by value, which will probably be changed in the future), but it works.
Things that this program can interpret:
- standard class declarations
- methods (with return types and arguments)
- arithmetic operators (+-*/), boolean operators (and/or)
- conditional if (though no comparisons are yet implemented, so it needs
an actual bool, and variable shadowing needs some work)
- variable declarations
- variable modification
- method calls
- printing
Arrays are not yet implemented.

Most programs that shouldn't get through the parser won't, but some
might, which will result in a runtime error like Exception:
Maybe.fromJust: Nothing.